// src/types/EventType.ts
export type EventType = {
  id: string;
  organizerType: string;
  organizerName: string;
  organizerEmail: string;
  phone: string;
  link: string;
  organizerPhoto: string;
  title: string;
  categoryId: string;
  opportunityType: string;
  assistanceType: string;
  target: string | number;
  region: string;
  address: string;
  startingDate: string;
  startHour: string;
  startMinute: string;
  endingDate: string;
  endHour: string;
  endMinute: string;
  startPeriod: string;
  endPeriod: string;
  timeDemands: string;
  skills: string;
  coverImage: string;
  coverUrl: string;
  description: string;
  descriptionLink: string;
  documentFile: string;
  currentProgress: string | number;
};